+++ web/src/main/java/swe/controller/IssueController.java
@@ -11,13 +11,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import swe.dto.CommentAddRequest;
-import swe.dto.IssueDescriptionUpdateRequest;
-import swe.dto.IssuePriorityNameResponse;
-import swe.dto.IssuePriorityUpdateRequest;
-import swe.dto.IssueResponse;
-import swe.dto.IssueStatusNameResponse;
-import swe.dto.IssueStatusUpdateRequest;
+import swe.dto.issue.CommentAddRequest;
+import swe.dto.issue.IssueDescriptionUpdateRequest;
+import swe.dto.issue.IssuePriorityNameResponse;
+import swe.dto.issue.IssuePriorityUpdateRequest;
+import swe.dto.issue.IssueResponse;
+import swe.dto.issue.IssueStatusNameResponse;
+import swe.dto.issue.IssueStatusUpdateRequest;
 import swe.issue.application.IssueService;
 import swe.issue.dto.IssueCreateRequest;
 import swe.user.dto.JwtMemberId;
