+++ web/src/main/java/swe/controller/UserController.java
@@ -6,11 +6,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import swe.dto.DuplicateNicknameRequest;
-import swe.dto.DuplicateNicknameResponse;
-import swe.dto.LoginResponse;
-import swe.dto.UserLoginRequest;
-import swe.dto.UserRoleResponse;
+import swe.dto.user.DuplicateNicknameRequest;
+import swe.dto.user.DuplicateNicknameResponse;
+import swe.dto.user.LoginResponse;
+import swe.dto.user.UserLoginRequest;
+import swe.dto.user.UserRoleResponse;
 import swe.user.application.JwtProvider;
 import swe.user.application.UserService;
 import swe.user.dto.UserRegisterRequest;
