+++ web/src/main/java/swe/dto/issue/IssueDescriptionUpdateRequest.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.issue;
 
 public record IssueDescriptionUpdateRequest(String newDescription) {
