+++ web/src/main/java/swe/dto/issue/IssuePriorityUpdateRequest.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.issue;
 
 public record IssuePriorityUpdateRequest(String priority) {
