+++ web/src/main/java/swe/dto/issue/IssueResponse.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.issue;
 
 import java.util.List;
 import swe.issue.domain.Issue;
