+++ web/src/main/java/swe/dto/issue/IssueStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.issue;
 
 public record IssueStatusUpdateRequest(String status) {
