+++ web/src/main/java/swe/dto/project/ProjectCreateRequest.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.project;
 
 public record ProjectCreateRequest(String title) {
