+++ web/src/main/java/swe/dto/user/DuplicateNicknameResponse.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.user;
 
 public record DuplicateNicknameResponse(Boolean isDuplicated) {
