+++ web/src/main/java/swe/dto/user/LoginResponse.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.user;
 
 public record LoginResponse(String token) {
