+++ web/src/main/java/swe/dto/user/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package swe.dto;
+package swe.dto.user;
 
 public record UserLoginRequest(String accountId, String password) {
